Move some classes. Create separate controllers
