+++ src/main/java/com/zemlyak/web/segmentapp/segment/CountryStatsUpdateRepository.java
@@ -1,6 +1,5 @@
-package com.zemlyak.web.segmentapp;
+package com.zemlyak.web.segmentapp.segment;
 
-import com.zemlyak.web.segmentapp.model2.CountryStat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CountryStatsUpdateRepository extends JpaRepository<CountryStat, CountryStat.Key> {
