+++ src/main/java/com/zemlyak/web/segmentapp/stat/Segment.java
@@ -1,36 +1,48 @@
-package com.zemlyak.web.segmentapp.model;
+package com.zemlyak.web.segmentapp.stat;
 
+import com.zemlyak.web.segmentapp.common.DataProvider;
+import com.zemlyak.web.segmentapp.common.SegmentType;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.proxy.HibernateProxy;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.util.Objects;
 
 @Getter
 @Setter
 @ToString
-@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
-@Table(name = "data_providers")
-public class DataProvider {
+@Table(name = "segments")
+public class Segment {
     @Id
     @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Integer id;
 
     @Column(name = "name")
     private String name;
 
-    public DataProvider(Integer id, String name) {
-        this();
-        this.id = id;
-        this.name = name;
-    }
+    @Column(name = "is_active")
+    private Boolean active;
+
+    @ManyToOne
+    @JoinColumn(name = "data_provider")
+    private DataProvider dataProvider;
+
+    @ManyToOne
+    @JoinColumn(name = "type_id")
+    private SegmentType segmentType;
 
     @Override
     public final boolean equals(Object o) {
@@ -39,8 +51,8 @@ public class DataProvider {
         Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        DataProvider that = (DataProvider) o;
-        return getId() != null && Objects.equals(getId(), that.getId());
+        Segment segment = (Segment) o;
+        return getId() != null && Objects.equals(getId(), segment.getId());
     }
 
     @Override
