+++ src/main/java/com/zemlyak/web/segmentapp/tracking/TrackingSetting.java
@@ -1,7 +1,7 @@
-package com.zemlyak.web.segmentapp.model;
+package com.zemlyak.web.segmentapp.tracking;
 
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.proxy.HibernateProxy;
@@ -9,47 +9,40 @@ import org.hibernate.proxy.HibernateProxy;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.InheritanceType;
 import jakarta.persistence.Table;
 import java.util.Objects;
 
 @Getter
 @Setter
 @ToString
-@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
-@Table(name = "segment_types")
-public class SegmentType {
+@Table(name = "tracking_settings")
+@Inheritance(strategy = InheritanceType.JOINED)
+public class TrackingSetting {
 
     @Id
     @Column(name = "id")
     private Integer id;
 
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "view_name")
-    private String viewName;
-
-    public SegmentType(Integer id, String name, String viewName) {
-        this();
-        this.id = id;
-        this.name = name;
-        this.viewName = viewName;
-    }
+    @Column(name = "metric_id")
+    private Integer metricId;
 
     @Override
-    public final boolean equals(Object o) {
+    public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null) return false;
         Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        SegmentType that = (SegmentType) o;
+        TrackingSetting that = (TrackingSetting) o;
         return getId() != null && Objects.equals(getId(), that.getId());
     }
 
     @Override
-    public final int hashCode() {
+    public int hashCode() {
         return this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass().hashCode() : getClass().hashCode();
     }
 }
